requirements.txt
playwright
python-dotenv
pandas
schedule
PyPDF2